third_party/WebKit/LayoutTests/fast/canvas/draw-system-focus-ring-on-event-expected.txt

PASS data[0] != 0 || data[1] != 0 || data[2] != 0 is true
PASS data[0] == 0 && data[1] == 0 && data[2] == 0 is true
PASS successfullyParsed is true

TEST COMPLETE
